refactor: split utility modules
