+++ instagrapi/mixins/album.py
@@ -9,7 +9,8 @@ from instagrapi.exceptions import (
     AlbumUnknownFormat,
 )
 from instagrapi.types import Location, Media, Track, Usertag
-from instagrapi.utils import date_time_original, dumps
+from instagrapi.utils.serialization import dumps
+from instagrapi.utils.timing import date_time_original
 
 
 class DownloadAlbumMixin:
