+++ instagrapi/mixins/hashtag.py
@@ -15,7 +15,7 @@ from instagrapi.extractors import (
     extract_media_v1,
 )
 from instagrapi.types import Hashtag, Media
-from instagrapi.utils import dumps
+from instagrapi.utils.serialization import dumps
 
 
 class HashtagMixin:
