+++ instagrapi/utils/auth.py
@@ -1,0 +1,32 @@
+import random
+import string
+import urllib.parse
+
+
+def generate_signature(data):
+    """Generate signature of POST data for Private API
+
+    Returns
+    -------
+    str
+        e.g. "signed_body=SIGNATURE.test"
+    """
+    return "signed_body=SIGNATURE.{data}".format(data=urllib.parse.quote_plus(data))
+
+
+def gen_token(size=10, symbols=False):
+    """Gen CSRF or something else token"""
+    chars = string.ascii_letters + string.digits
+    if symbols:
+        chars += string.punctuation
+    return "".join(random.choice(chars) for _ in range(size))
+
+
+def gen_password(size=10):
+    """Gen password"""
+    return gen_token(size)
+
+
+def generate_jazoest(symbols: str) -> str:
+    amount = sum(ord(s) for s in symbols)
+    return f"2{amount}"
