+++ instagrapi/utils/ids.py
@@ -1,0 +1,30 @@
+class InstagramIdCodec:
+    ENCODING_CHARS = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-_"
+
+    @staticmethod
+    def encode(num, alphabet=ENCODING_CHARS):
+        """Covert a numeric value to a shortcode."""
+        num = int(num)
+        if num == 0:
+            return alphabet[0]
+        arr = []
+        base = len(alphabet)
+        while num:
+            rem = num % base
+            num //= base
+            arr.append(alphabet[rem])
+        arr.reverse()
+        return "".join(arr)
+
+    @staticmethod
+    def decode(shortcode, alphabet=ENCODING_CHARS):
+        """Covert a shortcode to a numeric value."""
+        base = len(alphabet)
+        strlen = len(shortcode)
+        num = 0
+        idx = 0
+        for char in shortcode:
+            power = strlen - (idx + 1)
+            num += alphabet.index(char) * (base**power)
+            idx += 1
+        return num
