+++ instagrapi/utils/validation.py
@@ -1,0 +1,6 @@
+from instagrapi.exceptions import ValidationError
+
+
+def vassert(pred, message):
+    if not pred:
+        raise ValidationError(message)
